requirements.txt
snowflake-connector-python[pandas]
dotenv
